modified cues for how to create a pattern. Made the app unavailable for too small screens.
