+++ src/Header/SaveIcon.jsx
@@ -51,6 +51,7 @@ const S_SaveAlert = styled.div`
     position: absolute;
     right: 0px;
     width: 200px;
+    z-index: 1;
 `;
 
 const S_SaveIcon = styled.div`
