+++ src/Theme/S_DisplaySectionSubtitle.jsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const S_DisplaySectionSubtitle = styled.p`
+    font-size: 0.7rem;
+    font-variant: small-caps;
+    margin-top: 4px;
+    margin-bottom: 4px;
+`;
+
+export default S_DisplaySectionSubtitle;
