+++ src/Theme/S_DisplaySectionTitle.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const S_DisplaySectionTitle = styled.h2`
+    margin: ${({ marginTop }) => (marginTop ? "8" : "0")}px 0 0;
+`;
+
+export default S_DisplaySectionTitle;
