+++ src/Theme/constants.js
@@ -1,10 +1,13 @@
-const TITLE_MARGIN = 32;
-const TITLE_HEIGHT = 2;
+const GRID_MARGIN = 32;
+
+const MIN_WNDOW_HEIGHT = 600;
+const MIN_WNDOW_WIDTH = 900;
 
 const PATTERN_TITLE_MARGIN = 24;
 const PATTERN_TITLE_HEIGHT = 32;
 
-const GRID_MARGIN = 32;
+const TITLE_MARGIN = 32;
+const TITLE_HEIGHT = 2;
 
 export {
     TITLE_MARGIN,
@@ -12,4 +15,6 @@ export {
     PATTERN_TITLE_MARGIN,
     PATTERN_TITLE_HEIGHT,
     GRID_MARGIN,
+    MIN_WNDOW_HEIGHT,
+    MIN_WNDOW_WIDTH,
 };
